Refactor state specific page
